# api_rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto DAW Backend API",
        "description": "API REST del Proyecto DAW. Esta API gestiona la lógica de negocio para la plataforma de intercambio de tiempo.\n *     \n *     Módulos incluidos:\n *     - **Autenticación**: Registro, login, logout y gestión de perfil.\n *     - **Usuarios**: Gestión completa de usuarios (CRUD).\n *     - **Servicios**: Gestión de ofertas y demandas de servicios.\n *     - **Transacciones**: Control de intercambios de tiempo entre usuarios.\n *     - **Valoraciones**: Sistema de reputación y feedback.\n *     - **Mensajes**: Sistema de mensajería interna.\n *     - **Common**: Recursos compartidos (provincias, poblaciones, etc.).",
        "contact": {
            "email": "soporte@proyectodaw.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "https://proyecto-daw-backend.onrender.com/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/transacciones": {
            "get": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Obtener todas las transacciones",
                "operationId": "676ae9490a7cf488e1ae40645c11deea",
                "responses": {
                    "200": {
                        "description": "Lista de transacciones obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Todas las transacciones obtenidas correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transacciones/{usuario_id}": {
            "get": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Obtener transacciones de un usuario (como solicitante u ofertante)",
                "operationId": "d0f86bf3c6b6d45ace36f6472bb84077",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transacciones del usuario obtenidas correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transacciones del usuario obtenidas correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 406
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transaccion": {
            "post": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Crear una nueva transacción",
                "operationId": "a06c77f6f5580c8d9851fc9b42c73bc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "servicio_id",
                                    "usuario_solicitante_id",
                                    "usuario_ofertante_id",
                                    "horas"
                                ],
                                "properties": {
                                    "servicio_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usuario_solicitante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usuario_ofertante_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "horas": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "estado": {
                                        "description": "pendiente, confirmado, cancelado",
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "fecha_confirmacion": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transacción creada correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transaccion/{transaccion}": {
            "put": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Actualizar una transacción existente",
                "operationId": "79565dafa829b2b7b4643f5990666ffd",
                "parameters": [
                    {
                        "name": "transaccion",
                        "in": "path",
                        "description": "ID de la transacción",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "servicio_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usuario_solicitante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usuario_ofertante_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "horas": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "confirmado"
                                    },
                                    "fecha_confirmacion": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción actualizada correctamente"
                    },
                    "404": {
                        "description": "Transacción no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Eliminar una transacción",
                "operationId": "51993c807962effc7d37a29a227b5645",
                "parameters": [
                    {
                        "name": "transaccion",
                        "in": "path",
                        "description": "ID de la transacción",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transacción eliminada correctamente"
                    },
                    "404": {
                        "description": "Transacción no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación mediante token Laravel Sanctum. Ingrese el token directamente.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}